structure: migrate master data Api code to Presentation folder
