--- src/master-data/service/Api/Program.cs
@@ -1,35 +1,0 @@
-using MasterData.Api.Extensions;
-
-var builder = WebApplication.CreateBuilder(args);
-
-// Add services to the container.
-
-builder.SetupConfiguration();
-
-builder.Services.AddControllers();
-// Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
-builder.Services.AddEndpointsApiExplorer();
-builder.Services.AddSwaggerGen();
-
-builder.Services.AddCors(options =>
-{
-    options.AddPolicy("AllowAllPolicy", corsPolicyBuilder => corsPolicyBuilder.AllowAnyOrigin()
-    .AllowAnyMethod()
-    .AllowAnyHeader());
-});
-
-builder.Services.RegisterLayers();
-
-var app = builder.Build();
-
-app.UseSwaggerSwashbuckle();
-
-app.UseHttpsRedirection();
-
-app.UseAuthorization();
-
-app.UseCors("AllowAllPolicy");
-
-app.MapControllers();
-
-app.Run();
