+++ src/master-data/service/Presentation/Factories/GpsCoordinateFactory.cs
