+++ src/master-data/service/Presentation/MasterData.Api.csproj
@@ -1,0 +1,19 @@
+<Project Sdk="Microsoft.NET.Sdk.Web">
+
+  <PropertyGroup>
+    <TargetFramework>net10.0</TargetFramework>
+    <Nullable>enable</Nullable>
+    <ImplicitUsings>enable</ImplicitUsings>
+  </PropertyGroup>
+
+   <ItemGroup>
+    <PackageReference Include="Swashbuckle.AspNetCore" Version="10.0.1" />
+  </ItemGroup>
+
+   <ItemGroup>
+     <ProjectReference Include="..\Application\MasterData.Application.csproj" />
+     <ProjectReference Include="..\Client\MasterData.Client.csproj" />
+     <ProjectReference Include="..\Infrastructure\MasterData.Infrastructure.csproj" />
+   </ItemGroup>
+
+</Project>
