+++ src/master-data/service/Presentation/appsettings.json
@@ -1,0 +1,13 @@
+{
+  "Logging": {
+    "LogLevel": {
+      "Default": "Information",
+      "Microsoft.AspNetCore": "Warning"
+    }
+  },
+  "AllowedHosts": "*",
+  "MaximumCitiesBeforeWarning": 300,
+  "ConnectionStrings": {
+    "Postgres": "Host=localhost;Port=5432;Database=master_data;Username=postgres;Password=postgres;Ssl Mode=Disable;"
+  }
+}
